Fully working!

+++ Figura2/untitled/src/models/Shape.java
@@ -2,7 +2,7 @@ package models;
 
 public abstract class Shape {
 
-    abstract int area ();
-    abstract int perimeter ();
+    abstract double area();
+    abstract double perimeter();
 
 }
